cmr-api/app/auth.py
from fastapi import Depends, HTTPException, status
from fastapi.security import APIKeyHeader
import os
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("API_KEY", "development-api-key")
API_KEY_HEADER = APIKeyHeader(name="X-API-Key")

def get_api_key(api_key: str = Depends(API_KEY_HEADER)):
    if api_key == API_KEY:
        return api_key
    raise HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Invalid API Key",
    )
